chore(unity-bootstrap-theme): remove duplication of js files for static site
