+++ packages/unity-bootstrap-theme/src/js/data-layer.js
@@ -151,3 +151,7 @@
    */
   window.initDataLayer = window.initDataLayer || initDataLayer;
 })();
+
+/* Do Not Copy - Used for local development */
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.googleAnalytics = window.initDataLayer;
