--- packages/unity-bootstrap-theme/src/js/googleAnalytics.js
@@ -1,5 +1,0 @@
-require('./dataLayerExample');
-
-const googleAnalytics = window.initDataLayer;
-
-export { googleAnalytics };
