+++ packages/unity-bootstrap-theme/stories/atoms/alerts/alerts.examples.stories.js
@@ -1,6 +1,6 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
 export default createComponent("Alerts", "Atoms", "Examples");
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics.js";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const AlertsComponent = createStory(
   <div>
