+++ packages/unity-bootstrap-theme/stories/atoms/anchor-menu/anchor-menu.js
@@ -1,4 +1,4 @@
-import { googleAnalytics } from '../../../src/js/googleAnalytics';
+import { googleAnalytics } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const initializeAnchorMenu = () => {
   const globalHeader = document.getElementById('asu-header');
