+++ packages/unity-bootstrap-theme/stories/atoms/buttons/buttons.examples.stories.js
@@ -1,6 +1,6 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
 export default createComponent("Buttons", "Atoms", "Examples");
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const ButtonColorsComponent = createStory(
   <div>
