+++ packages/unity-bootstrap-theme/stories/atoms/buttons/buttons.templates.stories.js
@@ -1,5 +1,5 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics.js";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 /*
 We want to allow this component to change color, size, and disabled status.
