+++ packages/unity-bootstrap-theme/stories/atoms/form-fields/form-fields.examples.stories.js
@@ -1,6 +1,6 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
 export default createComponent("Form Fields", "Atoms", "Examples");
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics.js";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const TextInputs = createStory(
   <form class="uds-form">
