+++ packages/unity-bootstrap-theme/stories/atoms/sidebar/sidebar.examples.stories.js
@@ -1,6 +1,6 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
 export default createComponent("Sidebar", "Atoms", "Examples");
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics.js";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const Sidebar = createStory(
   <div class="container my-5">
