+++ packages/unity-bootstrap-theme/stories/molecules/cards/cards.examples.stories.js
@@ -3,7 +3,7 @@ import exampleImage from "./example-image.jpg";
 
 export default createComponent("Cards", "Molecules", "Examples");
 
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const cardImages = createStory(
   <div style={{ width: "50%" }}>
