+++ packages/unity-bootstrap-theme/stories/molecules/image-based-cards/image-based-cards.templates.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createComponent, createStory } from '../../../helpers/wrapper.js';
-import { googleAnalytics as initFunc } from '../../../src/js/googleAnalytics';
+import { googleAnalytics as initFunc } from '@asu/unity-bootstrap-theme/js/data-layer.js';
 import portraitImage from './portrait-image.jpg';
 
 export default createComponent('Image Based Cards', 'Molecules', 'Templates');
