+++ packages/unity-bootstrap-theme/stories/molecules/pagination/pagination.examples.stories.js
@@ -1,6 +1,6 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
 export default createComponent("Pagination", "Molecules", "Examples");
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics.js";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 
 export const DisabledAndActiveStatesComponent = createStory(
   <nav aria-label="...">
