+++ packages/unity-bootstrap-theme/stories/molecules/person-profile/person-profile.templates.stories.js
@@ -1,5 +1,5 @@
 import { createComponent, createStory } from "../../../helpers/wrapper.js";
-import { googleAnalytics as initFunc } from "../../../src/js/googleAnalytics";
+import { googleAnalytics as initFunc } from "@asu/unity-bootstrap-theme/js/data-layer.js";
 import { Button } from "../../../../components-core/src/components/Button";
 import anonPic from "./anon.png";
