+++ packages/unity-bootstrap-theme/stories/organisms/global-header/global-header.examples.stories.js
@@ -9,7 +9,7 @@ import {
   ScrolledState,
   Partner,
 } from "./global-header.templates.stories.js";
-import { initGlobalHeader as initFunc } from "./global-header.js";
+import { initGlobalHeader as initFunc } from "@asu/unity-bootstrap-theme/js/global-header.js";
 
 export const BasicExample = createStory(Basic, { initFunc });
