+++ server/views/asuheader/index.njk
@@ -8,7 +8,8 @@
   <!-- Bootstrap includes for asu-web-standards-bootstrap4-->
   <link rel="stylesheet" href="../@asu/unity-bootstrap-theme/css/unity-bootstrap-theme.bundle.css" type="text/css" media="all">
   <script type="text/javascript" src="./@asu/unity-bootstrap-theme/js/bootstrap.bundle.min.js"></script>
-  <script type="text/javascript" src="./@asu/unity-bootstrap-theme/js/global-header-datalayer.js"></script>j
+  <script type="text/javascript" src="./@asu/unity-bootstrap-theme/js/data-layer.js"></script>
+  <script type="text/javascript" src="./@asu/unity-bootstrap-theme/js/global-header.js"></script>
   <script src="https://kit.fontawesome.com/35fb12a3a9.js" crossorigin="anonymous"></script>
 
   <style>
@@ -738,6 +739,10 @@
       AsuCookieConsent.init();
     </script>
   </footer>
+  <script>
+    window.initDataLayer();
+    window.initGlobalHeader();
+  </script>
 </body>
 
 </html>
